src/requirements.txt
discord.py[voice]
PyGithub
simplejsondb